+++ tauri-app/src/lib/utils/raindexClient/getAllNetworks.ts
